updated migrations and seeds as per new readme

+++ data/migrations/20180905221136_recipe-ingredients.js
@@ -10,7 +10,7 @@ exports.up = function(knex, Promise) {
       .notNullable()
       .references('id')
       .inTable('ingredients');
-    table.integer('amount').notNullable();
+    table.float('amount').notNullable();
   });
 };
 exports.down = function(knex, Promise) {
